fix browser tab execution, allow to keep browsers opened after test execution for debug
