+++ src/executeTestPlan.js
@@ -36,11 +36,16 @@ export const executeTestPlan = async ({
   cooldownBetweenExecutions,
 
   maxExecutionsInParallel,
-
   completedExecutionLogAbbreviation = false,
   completedExecutionLogMerging = false,
   logSummary = true,
   updateProcessExitCode = true,
+  // stopAfterExecute: true to ensure runtime is stopped once executed
+  // because we have what we wants: execution is completed and
+  // we have associated coverage and capturedConsole
+  // passsing false means all node process and browsers launched stays opened
+  // (can eventually be used for debug)
+  stopAfterExecute = true,
 
   coverage = process.argv.includes("--cover") ||
     process.argv.includes("--coverage"),
@@ -145,6 +150,7 @@ export const executeTestPlan = async ({
 
     defaultMsAllocatedPerExecution,
     maxExecutionsInParallel,
+    stopAfterExecute,
     cooldownBetweenExecutions,
     completedExecutionLogMerging,
     completedExecutionLogAbbreviation,
