+++ test-manual/chromium-tab/file.html
@@ -1,0 +1,9 @@
+<html>
+  <head>
+    <meta charset="utf8" />
+    <link rel="icon" href="data:," />
+  </head>
+  <body>
+    <script type="module" src="./file.js"></script>
+  </body>
+</html>
